Move license file to OpenSim/Auxiliary/catch
