+++ OpenSim/Actuators/Test/testDeGrooteFregly2016Muscle.cpp
@@ -31,7 +31,7 @@
 #include <OpenSim/Actuators/CoordinateActuator.h>
 
 #define CATCH_CONFIG_MAIN
-#include <OpenSim/Auxiliary/catch.hpp>
+#include "OpenSim/Auxiliary/catch/catch.hpp"
 
 using namespace OpenSim;
