+++ OpenSim/Common/Test/testComponentInterface.cpp
@@ -34,8 +34,7 @@
 #include <simbody/internal/MobilizedBody_Ground.h>
 
 #define CATCH_CONFIG_MAIN
-#include <OpenSim/Auxiliary/catch.hpp>
-
+#include "OpenSim/Auxiliary/catch/catch.hpp"
 #include <random>
 
 namespace
