+++ OpenSim/Common/Test/testFunctions.cpp
@@ -30,7 +30,8 @@
 #include <OpenSim/Common/Sine.h>
 
 #define CATCH_CONFIG_MAIN
-#include <OpenSim/Auxiliary/catch.hpp>
+#include "OpenSim/Auxiliary/catch/catch.hpp"
+
 #include <OpenSim/Common/PolynomialFunction.h>
 
 using namespace OpenSim;
