+++ OpenSim/Common/Test/testLatinHypercubeDesign.cpp
@@ -24,7 +24,7 @@
 #include <OpenSim/Common/LatinHypercubeDesign.h>
 
 #define CATCH_CONFIG_MAIN
-#include <OpenSim/Auxiliary/catch.hpp>
+#include "OpenSim/Auxiliary/catch/catch.hpp"
 
 using namespace OpenSim;
