+++ OpenSim/Common/Test/testPath.cpp
@@ -25,7 +25,7 @@
 #include <OpenSim/Auxiliary/auxiliaryTestFunctions.h>
 
 #define CATCH_CONFIG_MAIN
-#include <OpenSim/Auxiliary/catch.hpp>
+#include "OpenSim/Auxiliary/catch/catch.hpp"
 
 /* The purpose of this test is strictly to check that classes derived from
  * the Path class work outside of the objects/components they are meant to
