+++ OpenSim/Common/Test/testSTOFileAdapter.cpp
@@ -27,7 +27,7 @@
 #include <unordered_set>
 
 #define CATCH_CONFIG_MAIN
-#include <OpenSim/Auxiliary/catch.hpp>
+#include "OpenSim/Auxiliary/catch/catch.hpp"
 
 using namespace OpenSim;
