+++ OpenSim/Common/Test/testStorage.cpp
@@ -25,8 +25,7 @@
 #include <OpenSim/Common/STOFileAdapter.h>
 
 #define CATCH_CONFIG_MAIN
-#include <OpenSim/Auxiliary/catch.hpp>
-
+#include "OpenSim/Auxiliary/catch/catch.hpp"
 #include <fstream>
 
 using namespace OpenSim;
