+++ OpenSim/Moco/Test/Testing.h
@@ -18,8 +18,9 @@
  * limitations under the License.                                             *
  * -------------------------------------------------------------------------- */
 
+#include "OpenSim/Auxiliary/catch/catch.hpp"
+
 #include <OpenSim/Common/osimCommon.h>
-#include <OpenSim/Auxiliary/catch.hpp>
 
 // Helper functions for comparing vectors.
 // ---------------------------------------
