+++ OpenSim/Simulation/Test/testPointToPointSpring.cpp
@@ -24,11 +24,11 @@
 #include <OpenSim/Simulation/Model/PointToPointSpring.h>
 
 #define CATCH_CONFIG_MAIN
-#include <OpenSim/Auxiliary/catch.hpp>
+#include "OpenSim/Auxiliary/catch/catch.hpp"
+#include <memory>
+
 #include <OpenSim/Simulation/Model/Model.h>
 #include <OpenSim/Simulation/SimbodyEngine/Body.h>
-
-#include <memory>
 
 // repro related to #3485
 //
