+++ OpenSim/Simulation/Test/testTableProcessor.cpp
@@ -17,7 +17,8 @@
  * -------------------------------------------------------------------------- */
 
 #define CATCH_CONFIG_MAIN
-#include <OpenSim/Auxiliary/catch.hpp>
+#include "OpenSim/Auxiliary/catch/catch.hpp"
+
 #include <OpenSim/Common/CommonUtilities.h>
 #include <OpenSim/Common/STOFileAdapter.h>
 #include <OpenSim/Simulation/TableProcessor.h>
