+++ OpenSim/Tests/Wrapping/testWrapping.cpp
@@ -28,7 +28,7 @@
 #include <SimTKcommon.h>
 
 #define CATCH_CONFIG_MAIN
-#include <OpenSim/Auxiliary/catch.hpp>
+#include "OpenSim/Auxiliary/catch/catch.hpp"
 
 using namespace OpenSim;
 using namespace SimTK;
