+++ Vendors/lepton/test/testLepton.cpp
@@ -23,7 +23,7 @@
 #include "Lepton.h"
 
 #define CATCH_CONFIG_MAIN
-#include <OpenSim/Auxiliary/catch.hpp>
+#include "OpenSim/Auxiliary/catch/catch.hpp"
 
 using namespace std;
