+++ Vendors/tropter/tests/test_analytic_optimal_control_solutions.cpp
@@ -14,7 +14,7 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------
 #define CATCH_CONFIG_MAIN
-#include <catch.hpp>
+#include "catch/catch.hpp"
 #include "testing.h"
 
 #include <tropter/tropter.h>
