+++ Vendors/tropter/tests/test_generic_optimization.cpp
@@ -16,11 +16,10 @@
 
 
 #define CATCH_CONFIG_MAIN
-#include <catch.hpp>
+#include "catch/catch.hpp"
+#include "testing.h"
 
 #include <tropter/tropter.h>
-
-#include "testing.h"
 
 using Eigen::Vector4d;
 using Eigen::Vector2d;
