+++ Vendors/tropter/tests/test_path_constraints.cpp
@@ -14,7 +14,7 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------
 #define CATCH_CONFIG_MAIN
-#include <catch.hpp>
+#include "catch/catch.hpp"
 #include "testing.h"
 
 #include <tropter/tropter.h>
