+++ Vendors/tropter/tests/test_sliding_mass_minimum_time.cpp
@@ -16,7 +16,7 @@
 #include <tropter/tropter.h>
 
 #define CATCH_CONFIG_MAIN
-#include <catch.hpp>
+#include "catch/catch.hpp"
 #include "testing_optimalcontrol.h"
 #include <cmath>
